[FIX]:Video 및 Channel id type 변경

+++ src/main/java/com/semtleWebGroup/youtubeclone/domain/channel/api/ChannelApi.java
@@ -1,7 +1,6 @@
 package com.semtleWebGroup.youtubeclone.domain.channel.api;
 
 import com.semtleWebGroup.youtubeclone.domain.channel.application.ChannelService;
-import com.semtleWebGroup.youtubeclone.domain.channel.application.SubscribeService;
 import com.semtleWebGroup.youtubeclone.domain.channel.domain.Channel;
 import com.semtleWebGroup.youtubeclone.domain.channel.dto.ChannelRequest;
 import lombok.RequiredArgsConstructor;
@@ -49,7 +48,7 @@ public class ChannelApi {
      * @return 현재는 entity 자체를 반환. 프론트 사용 데이터 보고 overfetching 줄일 예정
      */
     @GetMapping("/{channelId}")
-    public ResponseEntity getChannelInfo(@PathVariable("channelId")Long channelId){
+    public ResponseEntity getChannelInfo(@PathVariable("channelId")Integer channelId){
         Channel channel = channelService.getChannel(channelId);
 
         return ResponseEntity.status(HttpStatus.OK).body(channel);
@@ -110,7 +109,7 @@ public class ChannelApi {
      * @throws Exception 이미지 첨부 관련 Exception
      */
     @PatchMapping("/{channelId}")
-    public ResponseEntity editChannel(@PathVariable("channelId")Long channelId,
+    public ResponseEntity editChannel(@PathVariable("channelId")Integer channelId,
                                       @ModelAttribute ChannelRequest dto) throws Exception{
         final Channel channel = channelService.updateChannel(channelId, dto);
