+++ src/main/java/com/semtleWebGroup/youtubeclone/domain/video/service/VideoService.java
@@ -1,12 +1,10 @@
 package com.semtleWebGroup.youtubeclone.domain.video.service;
 
+import com.semtleWebGroup.youtubeclone.domain.video.domain.Video;
 import com.semtleWebGroup.youtubeclone.domain.video.dto.VideoRequest;
-import com.semtleWebGroup.youtubeclone.domain.video.entity.Video;
 import com.semtleWebGroup.youtubeclone.domain.video.repository.VideoRepository;
 import com.semtleWebGroup.youtubeclone.global.error.exception.EntityNotFoundException;
 import org.springframework.stereotype.Service;
-
-import java.util.NoSuchElementException;
 
 @Service
 public class VideoService {
@@ -16,7 +14,7 @@ public class VideoService {
         return new Video();
     }
 
-    public Video get(Long id) {
+    public Video get(Integer id) {
         Video video = videoRepository.findById(id)
                 .orElseThrow(()-> new EntityNotFoundException(
                         String.format("%d is not found.", id)
